Add files via upload
In this problem, you are given a csv le that contains the number of confirmed people with COVID-19 in a period of time in different countries. You are required to implement a recurrent neural network to classify if the number of confirmed people will increase in the next day.